+++ test/routes/BaseMapiNspiRoute.test.ts
@@ -15,10 +15,6 @@ import { BaseMapiNspiRoute } from "../../src/BaseMapiNspiRoute.js";
 class TestMapiNspiRoute extends BaseMapiNspiRoute<any> {
     protected mailboxClass: any = { name: "TestMailbox" };
     protected contactClass: any = { name: "TestContact" };
-
-    protected likePattern(term: string): string {
-        return term;
-    }
 }
 
 function makeReq(requestType?: string): any {